subdevice-model class to own file
